Add platform level errors. (#240)
In the presence of a Platform level error the
Scheduler will complete all the futures associated
to all the computations with a ShutdownError. That
means all the handles will raise such error when
`result()` is called on them.

A platform level error is an error that happens at
the server. At the time of this PR, the application
nodes assume the Server is always available and in
a good state, when that does not happen, the node
just tries to stop everything as graceful as
possible but with no gurantees.